Use post api with authenticate middleware

+++ src/services/PostService.ts
@@ -1,5 +1,6 @@
 import { Error } from 'mongoose'
 import PostModel, { Post } from '../models/Post'
+import { User } from '../models/User'
 
 type PostInput = Pick<Post, 'title' | 'content'>
 
@@ -18,19 +19,25 @@ export default class PostService {
     return post
   }
 
-  public async create(postInput: PostInput) {
-    const post = new PostModel(postInput)
-    const err = post.validateSync()
+  public async create(postInput: PostInput, user: User) {
+    const post = new PostModel({ ...postInput, author: user._id })
 
-    if (err instanceof Error.ValidationError) {
-      return null
+    try {
+      const err = post.validateSync()
+
+      if (err instanceof Error.ValidationError) {
+        return null
+      }
+
+      const res = await post.save()
+      return await res.populate({ path: 'author', select: ['_id', 'email'] }).execPopulate()
+    } catch (e) {
+      console.log(e)
     }
-
-    return await post.save()
   }
 
-  public async delete(id: string) {
-    const isDeleted = await PostModel.deleteOne({ _id: id })
+  public async delete(id: string, user: User) {
+    const isDeleted = await PostModel.deleteOne({ _id: id, author: user._id })
     return isDeleted.deletedCount
   }
 }
